# sipaling-kopi.swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sipaling Kopi API",
    "description": "API documentation for the Sipaling Kopi application.",
    "version": "1.0.0",
    "contact": {
      "email": "aidilbaihaqi8@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/admin": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Admin-only endpoint",
        "description": "An endpoint that can only be accessed by users with the 'admin' role.",
        "security": [
          {
            "sanctum": []
          },
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Admin access granted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/kitchen": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Kitchen Manager-only endpoint",
        "description": "An endpoint that can only be accessed by users with the 'kitchen_manager' role.",
        "security": [
          {
            "sanctum": []
          },
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Kitchen manager access granted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/cashier": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Cashier-only endpoint",
        "description": "An endpoint that can only be accessed by users with the 'cashier' role.",
        "security": [
          {
            "sanctum": []
          },
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cashier access granted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Laravel Sanctum authentication"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  }
}